test : Ball 판별 모델 테스트
